test: move load-time dependent test into separate file
